chore: optimisations
