+++ package.json
@@ -146,22 +146,22 @@
       "view/title": [
         {
           "command": "vsc-workspace-sidebar-collapse-all",
-          "group": "navigation",
+          "group": "navigation@0",
           "when": "view == vsc-workspace-sidebar-webview-workspace && vsc-workspace-sidebar-list-loaded && config.workspaceSidebar.showFolderHierarchy"
         },
         {
           "command": "vsc-workspace-sidebar-expand-all",
-          "group": "navigation",
+          "group": "navigation@1",
           "when": "view == vsc-workspace-sidebar-webview-workspace && vsc-workspace-sidebar-list-loaded && config.workspaceSidebar.showFolderHierarchy"
         },
         {
-          "command": "vsc-workspace-sidebar-refresh",
-          "group": "navigation",
-          "when": "view == vsc-workspace-sidebar-webview-workspace && vsc-workspace-sidebar-list-loaded"
+          "command": "vsc-workspace-sidebar-sort",
+          "group": "navigation@2",
+          "when": "view == vsc-workspace-sidebar-webview-workspace && vsc-workspace-sidebar-list-loaded && !config.workspaceSidebar.showFolderHierarchy"
         },
         {
-          "command": "vsc-workspace-sidebar-sort",
-          "group": "navigation",
+          "command": "vsc-workspace-sidebar-refresh",
+          "group": "navigation@3",
           "when": "view == vsc-workspace-sidebar-webview-workspace && vsc-workspace-sidebar-list-loaded"
         }
       ]
