+++ src/utils/arrays/index.ts
@@ -1,2 +1,2 @@
 export * from './findDuplicates';
-export * from './sortFilesByLabel';
+export * from './sortFilesByProp';
