+++ src/utils/arrays/sortFilesByProp.ts
@@ -1,0 +1,20 @@
+import { File } from '../../webviews/Workspace/WorkspaceViewProvider.interface';
+
+type SortFilesProps = 'label' | 'path';
+
+export const sortFilesByProp = (prop: SortFilesProps) => (a: File, b: File): number => {
+  if (prop === 'label' || prop === 'path') {
+    if (a.hasOwnProperty(prop) && b.hasOwnProperty(prop)) {
+      const aVal = String(a[prop]).toLowerCase();
+      const bVal = String(b[prop]).toLowerCase();
+
+      if (aVal > bVal) {
+        return 1;
+      } else if (aVal < bVal) {
+        return -1;
+      }
+    }
+  }
+
+  return 0;
+};
